% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.fragment.coordinates.R
\name{add.fragment.coordinates}
\alias{add.fragment.coordinates}
\title{add.fragment.coordinates}
\usage{
add.fragment.coordinates(id.data)
}
\arguments{
\item{id.data}{data table containing columns \code{target.id} and/or \code{bait.id} to be expanded}
}
\value{
Data table with added coordinate columns for target and bait (as applicable).
}
\description{
Expand target and bait IDs of the form chrN:start-end to separate coordinate columns in the data table
}
\examples{
data(bre80);
add.fragment.coordinates(bre80[, .(bait.id, target.id)]);

}
\author{
Erle Holgersen <Erle.Holgersen@icr.ac.uk>
}
