% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtools.installed.R
\name{bedtools.installed}
\alias{bedtools.installed}
\title{bedtools.installed}
\usage{
bedtools.installed()
}
\value{
Logical indicating if bedtools was found in PATH
}
\description{
Check if bedtools exists in PATH
}
\examples{
bedtools.installed();

}
\author{
Erle Holgersen <Erle.Holgersen@icr.ac.uk>
}
