% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.model.numerical.fit.R
\name{check.model.numerical.fit}
\alias{check.model.numerical.fit}
\title{check.model.identifiability}
\usage{
check.model.numerical.fit(interaction.data)
}
\arguments{
\item{interaction.data}{Data table of interaction data on which model is to be fit}
}
\value{
boolean indicating if model can be fit
}
\description{
Check if chicane model can be fit on a given dataset. 
\code{glm.nb} does not work when all responses are constant, or there are only two unique values and a covariate is a perfect predictor.
}
