% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.replicates.R
\name{compare.replicates}
\alias{compare.replicates}
\title{compare.replicates}
\usage{
compare.replicates(interaction.data = NULL, output.directory = "./")
}
\arguments{
\item{interaction.data}{A named vector specifying paths to files created using \{prepare.data()\}}

\item{output.directory}{Path to the output directory where pairwise plots are generated}
}
\value{
TRUE if pairwise plots were successfully created
}
\description{
Compare replicates in a pairwise manner and further stratified by distance
}
\examples{
# TODO

}
\author{
Syed Haider
}
