% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.bam.R
\name{convert.bam}
\alias{convert.bam}
\title{convert.bam}
\usage{
convert.bam(bam, baits, fragments, temp.directory = NULL, keep.files = FALSE)
}
\arguments{
\item{bam}{Path to a BAM file}

\item{baits}{Path to a BED file containing the baits}

\item{fragments}{Path to a BED file containing all restriction fragments in the genome}

\item{temp.directory}{Directory where temporary files should be stored. Defaults to current directory.}

\item{keep.files}{Logical indicating whether to keep temporary files}
}
\description{
Convert a BAM file to a format that can be used for replicate merging. 

	\strong{Note:} This function does not process data enough to be used for interaction calling. Use \code{prepare.data} for full preprocessing.
}
\seealso{
\code{\link{prepare.data}}
}
\author{
Erle Holgersen <Erle.Holgersen@icr.ac.uk>
}
