% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.hicup.digest.bed.R
\name{convert.hicup.digest.bed}
\alias{convert.hicup.digest.bed}
\title{convert.hicup.digest.bed}
\usage{
convert.hicup.digest.bed(hicup.digest, file.name = "")
}
\arguments{
\item{hicup.digest}{Path to HiCUP digest}

\item{file.name}{Path to output file. A blank string indicates output to the console.}
}
\description{
Convert a HiCUP digest file to BED format.
}
\examples{
hicup.digest <- system.file('extdata', 'HiCUP_digest_example.txt', package = 'chicane');
	convert.hicup.digest.bed(hicup.digest);

}
