% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.to.one.based.R
\name{convert.to.one.based}
\alias{convert.to.one.based}
\title{convert.to.one.based}
\usage{
convert.to.one.based(id)
}
\arguments{
\item{id}{string in format chr:start-end}
}
\value{
one-converted ID
}
\description{
Convert zero-based region in format chr:start-end to 1-based
}
