% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.modelfit.plot.R
\name{create.modelfit.plot}
\alias{create.modelfit.plot}
\title{create.modelfit.plot}
\usage{
create.modelfit.plot(model, file.name = NULL, resolution = 300)
}
\arguments{
\item{model}{An object of fitted model}

\item{file.name}{A string specifying plotting file name}

\item{resolution}{A numeric specifying plot's resolution}
}
\value{
TRUE if plot was successfully created
}
\description{
create a plot representing model's fit
}
\author{
Syed Haider
}
