% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.bin.R
\name{distance.bin}
\alias{distance.bin}
\title{distance.bin}
\usage{
distance.bin(distance)
}
\arguments{
\item{distance}{Vector of distances that should be mapped to a distance bin}
}
\value{
vector of same length as \code{distance} containing assigned distance bins
}
\description{
Assign distances to a meaningful category
}
