% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.in.zeros.R
\name{fill.in.zeros}
\alias{fill.in.zeros}
\alias{fill.in.zeroes}
\title{fill.in.zeros}
\usage{
fill.in.zeros(interaction.data, baits, fragments)

fill.in.zeroes(interaction.data, baits, fragments)
}
\arguments{
\item{interaction.data}{Data table containing interaction data}

\item{baits}{Vector of bait IDs used in the experiment, in format chrN:start-end}

\item{fragments}{Vector of potential fragments the baits can link up to, in format chrN:start-end}
}
\value{
Data table containing origiina
}
\description{
Add zero counts to interaction data
}
\examples{
data(bre80);
	bait.file <- system.file('extdata', '2q35.bed', package = 'chicane');
	fragment.file <- system.file('extdata', 'GRCh38_HindIII_chr2.bed.gz', package = 'chicane');
	results <- fill.in.zeros(
	bre80,
	baits = read.bed(bait.file), 
	fragments = read.bed(fragment.file)
	);

}
