% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.combination.count.R
\name{get.combination.count}
\alias{get.combination.count}
\title{get.combination.count}
\usage{
get.combination.count(baits, fragments, cis.only = FALSE)
}
\arguments{
\item{baits}{vector of bait IDs in form chrN:start-end}

\item{fragments}{vector of fragment IDs in form chrN:start-end}

\item{cis.only}{logical indicating whether cis-interactions only should be considered}
}
\value{
total number of possible combinations
}
\description{
Calculate the number of possible combinations between baits and fragments, 
	excluding self-ligations and only counting bait-to-bait interactions once (e.g. a-b, not b-a)
}
