% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.components.R
\name{get.components}
\alias{get.components}
\title{get.components}
\usage{
get.components(id)
}
\arguments{
\item{id}{string in format chr:start-end}
}
\value{
list with entries 'chr', 'start', 'end'
}
\description{
Split a fragment in format chr:start-end to a list of corresponding elements
}
