% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.distance.R
\name{get.distance}
\alias{get.distance}
\title{get.distance}
\usage{
get.distance(interaction.data)
}
\arguments{
\item{interaction.data}{data.table with interaction data. Must contain columns bait.chr, bait.start, bait.end, target.chr, target.start, target.end}
}
\value{
vector of absolute distances (NA for trans-interactions)
}
\description{
Calculate distance between bait and target region
}
\examples{
	data(bre80);
input.cols <- c('bait.chr', 'bait.start', 'bait.end', 
		'target.chr', 'target.start', 'target.end');
	get.distance( bre80[, input.cols, with = FALSE]);

}
