% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.components.R
\name{get.id.components}
\alias{get.id.components}
\title{get.id.components}
\usage{
get.id.components(id)
}
\arguments{
\item{id}{segment ID of form \code{chrN:start-end}}
}
\value{
A character vector of length three, where the elements are chromosome, start, and end, respectively. 
	If \code{id} is a vector, a list of the same length is returned
}
\description{
Split a segment ID in form \code{chrN:start-end} into its different components
}
\examples{
	get.id.components('chrX:6-30');
get.id.components(c('3:4-10', '22:1000-20000'))

}
