% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.interaction.id.R
\name{get.interaction.id}
\alias{get.interaction.id}
\title{get.interaction.id}
\usage{
get.interaction.id(bait, other.end, bait.to.bait, zero.based = FALSE)
}
\arguments{
\item{bait}{id of bait in format chr:start-end}

\item{other.end}{id of other end in format chr:start-end}

\item{bait.to.bait}{logical indicating whether both ends are baits}

\item{zero.based}{logical indicating if IDs are zero-based}
}
\value{
string identifying interaction
}
\description{
Generate a unique identifying ID for each interaction
}
