% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.glm.nb.maxiter.warning.R
\name{is.glm.nb.maxiter.warning}
\alias{is.glm.nb.maxiter.warning}
\title{is.glm.nb.maxiter.warning}
\usage{
is.glm.nb.maxiter.warning(w)
}
\arguments{
\item{w}{Warning object}
}
\value{
Logical indicating if warning matches iteration limit reached warning
}
\description{
Check if a warning object is an iteration limit reached warning from \code{glm.nb}
}
