% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.glm.nb.theta.error.R
\name{is.glm.nb.theta.error}
\alias{is.glm.nb.theta.error}
\title{check.glm.nb.theta.error}
\usage{
is.glm.nb.theta.error(e)
}
\arguments{
\item{e}{Error object}
}
\value{
Boolean indicating if error matches
}
\description{
Check if an error matches the error raised by \code{glm.nb} due to an inflated theta estimate.
This happens when the variance of the negative binomial does not exceed the mean (i.e. there is no overdispersion).
 In such cases, the Poisson distribution may be a suitable alternative.
}
