% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.glm.nb.theta.warning.R
\name{is.glm.nb.theta.warning}
\alias{is.glm.nb.theta.warning}
\title{is.glm.nb.theta.warning}
\usage{
is.glm.nb.theta.warning(w)
}
\arguments{
\item{w}{Warning object}
}
\value{
Boolean indicating if warning matches
}
\description{
Check if a warning matches the square root warning raised by \code{glm.nb} due to an inflated theta estimate.
This happens when the variance of the negative binomial does not exceed the mean (i.e. there is no overdispersion).
 In such cases, the Poisson distribution may be a suitable alternative.
}
