% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.modelfit.plot.R
\name{make.modelfit.plot}
\alias{make.modelfit.plot}
\title{make.modelfit.plot}
\usage{
make.modelfit.plot(model, file.name = NULL, resolution = 600)
}
\arguments{
\item{model}{An object of fitted model}

\item{file.name}{A string specifying plotting file name}

\item{resolution}{A numeric specifying plot's resolution}
}
\value{
TRUE if plot was successfully created
}
\description{
create a plot representing model's fit
}
\author{
Syed Haider <Syed.Haider@icr.ac.uk>
}
