% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.rows.sanity.check.R
\name{model.rows.sanity.check}
\alias{model.rows.sanity.check}
\title{model.rows.sanity.check}
\usage{
model.rows.sanity.check(model.data, model)
}
\arguments{
\item{model.data}{Data used to fit model}

\item{model}{Resulting negative binomial model object}
}
\value{
None
}
\description{
Check that the model fit contains the same number of rows as the data used to fit it, 
and throw an error if not
}
