% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bed.R
\name{read.bed}
\alias{read.bed}
\title{read.bed}
\usage{
read.bed(bed.path, zero.based = TRUE)
}
\arguments{
\item{bed.path}{Path to bed file}

\item{zero.based}{Whether to return ID in zero-based coordinates}
}
\value{
vector of region IDs
}
\description{
Read a BED file and return regions in chrN:start-end format
}
\examples{
bait.file <- system.file('extdata', '2q35.bed', package = 'chicane');
baits <- read.bed(bait.file);

}
