% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart.split.R
\name{smart.split}
\alias{smart.split}
\title{smart.split}
\usage{
smart.split(dat, bins)
}
\arguments{
\item{dat}{Data frame or data table to be split}

\item{bins}{Number of bins to split data into}
}
\value{
List with \code{bins} elements. Each element corresponds to one portion 
of the data
}
\description{
Split a data frame into a prespecified number of bins, using 
\code{split} and \code{cut}. Unlike the default R functions, this does not
fail when asked to split the data into a single bin.
}
