% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.enrichment.sample.R
\name{stratified.enrichment.sample}
\alias{stratified.enrichment.sample}
\title{stratified.enrichment.sample}
\usage{
stratified.enrichment.sample(nonsignificant.results, significant.results)
}
\arguments{
\item{nonsignificant.results}{Data table containing non-significant interactions that should be sampled from}

\item{significant.results}{Data table of significant results. Used to determine size of strata in stratified sampling procedure.}
}
\description{
Generate a stratified sample matching distance distribution of significant interactions.
}
