% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.interaction.data.R
\name{verify.interaction.data}
\alias{verify.interaction.data}
\title{verify.interaction.data}
\usage{
verify.interaction.data(interaction.data)
}
\arguments{
\item{interaction.data}{Object to be verified.}
}
\value{
None
}
\description{
Verify that interaction.data object is in expected format. Throws an error if object does not fit requirements.
}
